Create Path planning for robots
The proposed system aims at designing an autonomous, an automatic and an intelligent navigating robot. Hence a powerful platform such as Arduino a hardware independent interface is made use of. It is used as the brain of this project. It eases
interactions between the robot and the software platform.

Wireless camera is used in order to capture video stream and as well to make it to be independent with the base station and store the entire video to create the path planning of the vehicle in the dynamic video. ZigBee is used so as to have wireless communication between the robot and the base station, this technology is far better than the Bluetooth in its range of
area coverage.

This project makes use of efficient software named MATLAB R2013a in order to process the camera captured video stream. It uses effective and vital toolboxes and forms the heart of this system. It is used to find the distance of the obstacle in
the snapshot captured. It is used to find the depth information of the snapshot to confirm the existence of the obstacle in the snapshot. As well as, to provoke the speech using the base station (default speaker) and mainly to simulate the shortest path of the vehicle for its navigation in an environment.